+++ eras/alonzo/impl/testlib/Test/Cardano/Ledger/Alonzo/Arbitrary.hs
@@ -372,6 +372,9 @@ deriving instance Arbitrary ix => Arbitrary (AsIndex ix it)
 
 deriving instance Arbitrary it => Arbitrary (AsItem ix it)
 
+instance (Arbitrary ix, Arbitrary it) => Arbitrary (AsIxItem ix it) where
+  arbitrary = AsIxItem <$> arbitrary <*> arbitrary
+
 genAlonzoPlutusPurposePointer :: Word32 -> Gen (AlonzoPlutusPurpose AsIndex era)
 genAlonzoPlutusPurposePointer i =
   elements
@@ -386,6 +389,20 @@ instance
   , Arbitrary (TxCert era)
   ) =>
   Arbitrary (AlonzoPlutusPurpose AsItem era)
+  where
+  arbitrary =
+    oneof
+      [ AlonzoSpending <$> arbitrary
+      , AlonzoMinting <$> arbitrary
+      , AlonzoCertifying <$> arbitrary
+      , AlonzoRewarding <$> arbitrary
+      ]
+
+instance
+  ( Era era
+  , Arbitrary (TxCert era)
+  ) =>
+  Arbitrary (AlonzoPlutusPurpose AsIxItem era)
   where
   arbitrary =
     oneof