+++ eras/conway/impl/src/Cardano/Ledger/Conway/Core.hs
@@ -1,3 +1,5 @@
+{-# LANGUAGE PatternSynonyms #-}
+
 module Cardano.Ledger.Conway.Core (
   ConwayEraTxBody (..),
   ConwayEraPParams,
@@ -24,6 +26,9 @@ module Cardano.Ledger.Conway.Core (
   dvtPPTechnicalGroupL,
   dvtPPEconomicGroupL,
   dvtUpdateToConstitutionL,
+  ConwayEraScript (..),
+  pattern VotingPurpose,
+  pattern ProposingPurpose,
   module Cardano.Ledger.Babbage.Core,
 )
 where
@@ -54,6 +59,11 @@ import Cardano.Ledger.Conway.PParams (
   ppuGovActionDepositL,
   ppuGovActionLifetimeL,
   ppuPoolVotingThresholdsL,
+ )
+import Cardano.Ledger.Conway.Scripts (
+  ConwayEraScript (..),
+  pattern ProposingPurpose,
+  pattern VotingPurpose,
  )
 import Cardano.Ledger.Conway.Tx ()
 import Cardano.Ledger.Conway.TxBody (ConwayEraTxBody (..))